services/delivery_service/pyproject.toml
[tool.poetry]
name = "delivery-service"
version = "0.1.0"
description = ""
authors = ["Михаил <58465577+bubled33@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.123.10"
uvicorn = "^0.38.0"
pydantic-settings = "^2.12.0"
asyncpg = "^0.31.0"
libs = {path = "../../libs", develop = true}
pytest = "^9.0.2"
httpx = "^0.28.1"
prometheus-client = "^0.23.1"
yoyo-migrations = "^9.0.0"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.3.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
